% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{config}
\alias{config}
\alias{setTrnglThreads}
\alias{getTrnglThreads}
\alias{setTrnglSeed}
\alias{getTrnglSeed}
\title{trngl configuration functions}
\usage{
setTrnglThreads(n_threads)

getTrnglThreads()

setTrnglSeed(seed)

getTrnglSeed()
}
\arguments{
\item{n_threads}{Number of threads to use}

\item{seed}{Seed for the random number generator}
}
\description{
trngl uses an internal threadsafe random number generator. These
functions can be used to configure it.
}
