% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trngl.R
\name{convert}
\alias{convert}
\alias{cum2incr}
\alias{incr2cum}
\title{Convert between representations of claims triangles}
\usage{
cum2incr(trngl)

incr2cum(trngl)
}
\arguments{
\item{trngl}{For \code{cum2incr} a cumulative claims triangle, for \code{incr2cum} an incremental one.}
}
\value{
\code{cum2incr} returns the cumulative representation of a cumulative claims triangle.
Vice versa for \code{incr2cum}.
}
\description{
Convert between representations of claims triangles
}
\examples{
UKMotor.incr <- cum2incr(UKMotor)
all.equal(UKMotor, incr2cum(UKMotor.incr))
}
