% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.R
\name{mack}
\alias{mack}
\alias{mackParamBoot}
\alias{mackResidBoot}
\alias{mackPairsBoot}
\alias{mackPairsSim}
\alias{mackParamSim}
\alias{mackResidSim}
\title{Bootstrapping and simulation functions for the Mack chain ladder model}
\usage{
mackParamBoot(trngl, dist, cond, n_boot = 1000, n_sim = 10,
  progress = TRUE)

mackResidBoot(trngl, resid_type, cond, n_boot = 1000, n_sim = 10,
  progress = TRUE)

mackPairsBoot(trngl, n_boot = 1000, n_sim = 10, progress = TRUE)

mackPairsSim(trngl, sim_type, n_boot = 1000, n_sim = 10, progress = TRUE)

mackParamSim(trngl, sim_type, cond, dist, n_boot = 1000, n_sim = 10,
  progress = TRUE)

mackResidSim(trngl, sim_type, cond, resid_type, n_boot = 1000, n_sim = 10,
  progress = TRUE)
}
\arguments{
\item{trngl}{A \emph{cumulative} claims triangle}

\item{dist}{Distribution from which to simulate: "normal", "gamma" or "poisson"}

\item{cond}{Whether the bootstrap should be conditional}

\item{n_boot}{Number of bootstrap iterations}

\item{n_sim}{Number of simulation iterations}

\item{progress}{Whether to show progress}

\item{resid_type}{Which type of residuals to use: "standardised", "studentised" or "log-normal"}

\item{sim_type}{Type of simulation: "single", "origin" or "calendar"}
}
\value{
An object of class 'mack' containing the simulation results. There are S3 methods
for printing and plotting it. The object also inherits from a class depending on the value
of \code{sim_type}.
}
\description{
Bootstrapping and simulation functions for the Mack chain ladder model
}
\examples{
res <- mackParamSim(UKMotor, "single",
  cond = TRUE,
  dist = "normal",
  n_boot = 1e2,
  n_sim = 1e2,
  progress = FALSE
)
print(res)
}
