% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odp.R
\name{odp}
\alias{odp}
\alias{odpParamBoot}
\alias{odpResidBoot}
\alias{odpParamSim}
\alias{odpResidSim}
\title{Bootstrapping and simulation functions for the overdispersed Poisson model}
\usage{
odpParamBoot(trngl, dist, n_boot = 1000, n_sim = 1000, progress = TRUE)

odpResidBoot(trngl, n_boot = 1000, n_sim = 1000, progress = TRUE)

odpParamSim(trngl, sim_type, dist, n_boot = 1000, n_sim = 1000,
  progress = TRUE)

odpResidSim(trngl, sim_type, n_boot = 1000, n_sim = 1000,
  progress = TRUE)
}
\arguments{
\item{trngl}{A \emph{cumulative} claims triangle}

\item{dist}{Distribution from which to simulate: "normal", "gamma" or "poisson"}

\item{n_boot}{Number of bootstrap iterations}

\item{n_sim}{Number of simulation iterations}

\item{progress}{Whether to show progress}

\item{sim_type}{Type of simulation: "single", "origin" or "calendar"}
}
\value{
An object of class 'odp' containing the simulation results. There are S3 methods
for printing and plotting it. The object also inherits from a class depending on the value
of \code{sim_type}.
}
\description{
Bootstrapping and simulation functions for the overdispersed Poisson model
}
\examples{
res <- odpParamSim(UKMotor, "single",
  dist = "normal",
  n_boot = 1e2,
  n_sim = 1e2,
  progress = FALSE
)
print(res)
}
