% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trngl-package.R
\docType{package}
\name{trngl}
\alias{trngl}
\alias{trngl-package}
\title{trngl}
\description{
Bootstrap simulation for actuarial reserving models.
}
\author{
Othman El Hammouchi \href{mailto:othman.el.hammouchi@protonmail.com}{othman.el.hammouchi@protonmail.com}
}
\keyword{internal}
