% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trngl-package.R
\docType{data}
\name{UKMotor}
\alias{UKMotor}
\title{UK motor claims triangle}
\format{
An object of class \code{trngl} (inherits from \code{triangle}, \code{matrix}) with 7 rows and 7 columns.
}
\usage{
UKMotor
}
\description{
Re-export of \link[ChainLadder]{UKMotor}
}
\keyword{datasets}
